Move trait and implement in respective crates
