+++ epoch-rewards/Cargo.toml
@@ -23,7 +23,7 @@ frozen-abi = [
 ]
 serde = ["dep:serde", "dep:serde_derive", "serde/alloc", "solana-hash/serde"]
 std = ["serde?/std"]
-sysvar = ["dep:solana-sdk-ids", "dep:solana-sysvar-id"]
+sysvar = ["dep:solana-get-sysvar", "dep:solana-sdk-ids", "dep:solana-sysvar-id"]
 wincode = ["dep:wincode", "solana-hash/wincode"]
 
 [dependencies]
@@ -31,6 +31,7 @@ serde = { workspace = true, optional = true }
 serde_derive = { workspace = true, optional = true }
 solana-frozen-abi = { workspace = true, optional = true, features = ["frozen-abi"] }
 solana-frozen-abi-macro = { workspace = true, optional = true }
+solana-get-sysvar = { workspace = true, optional = true }
 solana-hash = { workspace = true, features = ["decode"] }
 solana-sdk-ids = { workspace = true, optional = true }
 solana-sdk-macro = { workspace = true }
