+++ epoch-schedule/Cargo.toml
@@ -15,7 +15,12 @@ targets = ["x86_64-unknown-linux-gnu"]
 [features]
 frozen-abi = ["dep:solana-frozen-abi", "dep:solana-frozen-abi-macro"]
 serde = ["dep:serde", "dep:serde_derive"]
-sysvar = ["dep:solana-sdk-ids", "dep:solana-sysvar-id"]
+sysvar = [
+    "dep:solana-get-sysvar",
+    "dep:solana-program-error",
+    "dep:solana-sdk-ids",
+    "dep:solana-sysvar-id",
+]
 wincode = ["dep:wincode"]
 
 [dependencies]
@@ -23,6 +28,8 @@ serde = { workspace = true, optional = true }
 serde_derive = { workspace = true, optional = true }
 solana-frozen-abi = { workspace = true, optional = true, features = ["frozen-abi"] }
 solana-frozen-abi-macro = { workspace = true, optional = true }
+solana-get-sysvar = { workspace = true, optional = true }
+solana-program-error = { workspace = true, optional = true }
 solana-sdk-ids = { workspace = true, optional = true }
 solana-sdk-macro = { workspace = true }
 solana-sysvar-id = { workspace = true, optional = true }
