+++ last-restart-slot/src/sysvar.rs
@@ -1,4 +1,12 @@
 pub use solana_sdk_ids::sysvar::last_restart_slot::{check_id, id, ID};
-use {crate::LastRestartSlot, solana_sysvar_id::impl_sysvar_id};
+use {
+    crate::LastRestartSlot,
+    solana_get_sysvar::{impl_sysvar_get, GetSysvar},
+    solana_sysvar_id::impl_sysvar_id,
+};
 
 impl_sysvar_id!(LastRestartSlot);
+
+impl GetSysvar for LastRestartSlot {
+    impl_sysvar_get!(id());
+}
