+++ rent/src/sysvar.rs
@@ -1,4 +1,12 @@
 pub use solana_sdk_ids::sysvar::rent::{check_id, id, ID};
-use {crate::Rent, solana_sysvar_id::impl_sysvar_id};
+use {
+    crate::Rent,
+    solana_get_sysvar::{impl_sysvar_get, GetSysvar},
+    solana_sysvar_id::impl_sysvar_id,
+};
 
 impl_sysvar_id!(Rent);
+
+impl GetSysvar for Rent {
+    impl_sysvar_get!(id(), 7);
+}
