+++ slot-hashes/src/sysvar.rs
@@ -1,4 +1,6 @@
 pub use solana_sdk_ids::sysvar::slot_hashes::{check_id, id, ID};
-use {crate::SlotHashes, solana_sysvar_id::impl_sysvar_id};
+use {crate::SlotHashes, solana_get_sysvar::GetSysvar, solana_sysvar_id::impl_sysvar_id};
 
 impl_sysvar_id!(SlotHashes);
+
+impl GetSysvar for SlotHashes {}
