+++ slot-history/src/sysvar.rs
@@ -1,4 +1,6 @@
 pub use solana_sdk_ids::sysvar::slot_history::{check_id, id, ID};
-use {crate::SlotHistory, solana_sysvar_id::impl_sysvar_id};
+use {crate::SlotHistory, solana_get_sysvar::GetSysvar, solana_sysvar_id::impl_sysvar_id};
 
 impl_sysvar_id!(SlotHistory);
+
+impl GetSysvar for SlotHistory {}
